+++ src/connectors/numeric-menu/__tests__/connectNumericMenu-test.ts
@@ -1128,7 +1128,8 @@ See documentation: https://www.algolia.com/doc/api-reference/widgets/numeric-men
   });
 
   describe('insights', () => {
-    it('sends event when a facet is added', () => {
+    // See: https://github.com/algolia/instantsearch.js/pull/5085
+    it(`doesn't send event when a facet is added`, () => {
       const rendering = jest.fn();
       const makeWidget = connectNumericMenu(rendering);
       const widget = makeWidget({
@@ -1143,47 +1144,15 @@ See documentation: https://www.algolia.com/doc/api-reference/widgets/numeric-men
       });
 
       const helper = jsHelper(createSearchClient(), '');
-      helper.search = jest.fn();
-      const initOptions = createInitOptions({
-        helper,
-        state: helper.state,
-      });
+      const initOptions = createInitOptions({ helper, state: helper.state });
       const { instantSearchInstance } = initOptions;
       widget.init!(initOptions);
 
       const firstRenderingOptions = rendering.mock.calls[0][0];
       const { refine, items } = firstRenderingOptions;
       refine(items[0].value);
-      expect(instantSearchInstance.sendEventToInsights).toHaveBeenCalledTimes(
-        1
-      );
-      expect(instantSearchInstance.sendEventToInsights).toHaveBeenCalledWith({
-        attribute: 'numerics',
-        eventType: 'click',
-        insightsMethod: 'clickedFilters',
-        payload: {
-          eventName: 'Filter Applied',
-          filters: ['numerics<=10'],
-          index: '',
-        },
-        widgetType: 'ais.numericMenu',
-      });
 
-      refine(items[1].value);
-      expect(instantSearchInstance.sendEventToInsights).toHaveBeenCalledTimes(
-        2
-      );
-      expect(instantSearchInstance.sendEventToInsights).toHaveBeenCalledWith({
-        attribute: 'numerics',
-        eventType: 'click',
-        insightsMethod: 'clickedFilters',
-        payload: {
-          eventName: 'Filter Applied',
-          filters: ['numerics<=20', 'numerics>=10'],
-          index: '',
-        },
-        widgetType: 'ais.numericMenu',
-      });
+      expect(instantSearchInstance.sendEventToInsights).not.toHaveBeenCalled();
     });
   });
 
@@ -1200,6 +1169,7 @@ See documentation: https://www.algolia.com/doc/api-reference/widgets/numeric-men
         numerics: {
           createURL: expect.any(Function),
           hasNoResults: true,
+          canRefine: false,
           items: [
             {
               isRefined: false,
@@ -1259,6 +1229,7 @@ See documentation: https://www.algolia.com/doc/api-reference/widgets/numeric-men
           refine: renderState1.numericMenu.numerics.refine,
           sendEvent: renderState1.numericMenu.numerics.sendEvent,
           hasNoResults: true,
+          canRefine: false,
           items: [
             {
               isRefined: false,
@@ -1340,6 +1311,7 @@ See documentation: https://www.algolia.com/doc/api-reference/widgets/numeric-men
         refine: expect.any(Function),
         sendEvent: expect.any(Function),
         hasNoResults: true,
+        canRefine: false,
         widgetParams: {
           attribute: 'numerics',
           items: [
@@ -1390,6 +1362,7 @@ See documentation: https://www.algolia.com/doc/api-reference/widgets/numeric-men
         refine: expect.any(Function),
         sendEvent: expect.any(Function),
         hasNoResults: true,
+        canRefine: false,
         widgetParams: {
           attribute: 'numerics',
           items: [
@@ -1425,6 +1398,7 @@ See documentation: https://www.algolia.com/doc/api-reference/widgets/numeric-men
       expect(renderState2).toEqual({
         createURL: expect.any(Function),
         hasNoResults: true,
+        canRefine: false,
         items: [
           {
             isRefined: false,
@@ -1462,6 +1436,28 @@ See documentation: https://www.algolia.com/doc/api-reference/widgets/numeric-men
             },
           ],
         },
+      });
+    });
+
+    describe('canRefine', () => {
+      it('should be true if there are no results but a refinement is being applied', () => {
+        const [widget, helper] = getInitializedWidget();
+        helper.setQueryParameter('numericRefinements', {
+          numerics: { '>=': [20] },
+        });
+        const renderState = widget.getWidgetRenderState(
+          createInitOptions({ state: helper.state, helper })
+        );
+        expect(renderState.canRefine).toBe(true);
+      });
+
+      it('should be false if there are no results and no refinement is being applied', () => {
+        const [widget, helper] = getInitializedWidget();
+
+        const renderState = widget.getWidgetRenderState(
+          createInitOptions({ state: helper.state, helper })
+        );
+        expect(renderState.canRefine).toBe(false);
       });
     });
   });