+++ src/connectors/numeric-menu/connectNumericMenu.ts
@@ -3,7 +3,6 @@ import {
   checkRendering,
   createDocumentationMessageGenerator,
   isFiniteNumber,
-  convertNumericRefinementsToFilters,
   noop,
 } from '../../lib/utils';
 import type {
@@ -13,10 +12,7 @@ import type {
   TransformItems,
   WidgetRenderState,
 } from '../../types';
-import type {
-  AlgoliaSearchHelper,
-  SearchParameters,
-} from 'algoliasearch-helper';
+import type { SearchParameters } from 'algoliasearch-helper';
 import type { InsightsEvent } from '../../middlewares';
 
 const withUsage = createDocumentationMessageGenerator({
@@ -91,8 +87,17 @@ export type NumericMenuRenderState = {
 
   /**
    * `true` if the last search contains no result
+   * @deprecated Use `canRefine` instead.
    */
   hasNoResults: boolean;
+
+  /**
+   * Indicates if search state can be refined.
+   *
+   * This is `true` if the last search contains no result and
+   * "All" range is selected
+   */
+  canRefine: boolean;
 
   /**
    * Sets the selected value and trigger a new search
@@ -132,45 +137,11 @@ export type NumericMenuConnector = Connector<
 const $$type = 'ais.numericMenu';
 
 const createSendEvent =
-  ({
-    instantSearchInstance,
-    helper,
-    attribute,
-  }: {
-    instantSearchInstance: InstantSearch;
-    helper: AlgoliaSearchHelper;
-    attribute: string;
-  }) =>
+  ({ instantSearchInstance }: { instantSearchInstance: InstantSearch }) =>
   (...args: [InsightsEvent] | [string, string, string?]) => {
     if (args.length === 1) {
       instantSearchInstance.sendEventToInsights(args[0]);
       return;
-    }
-
-    const [eventType, facetValue, eventName = 'Filter Applied'] = args;
-    if (eventType !== 'click') {
-      return;
-    }
-    // facetValue === "%7B%22start%22:5,%22end%22:10%7D"
-    const filters = convertNumericRefinementsToFilters(
-      getRefinedState(helper.state, attribute, facetValue),
-      attribute
-    );
-    if (filters && filters.length > 0) {
-      /*
-        filters === ["price<=10", "price>=5"]
-      */
-      instantSearchInstance.sendEventToInsights({
-        insightsMethod: 'clickedFilters',
-        widgetType: $$type,
-        eventType,
-        payload: {
-          eventName,
-          index: helper.getIndex(),
-          filters,
-        },
-        attribute,
-      });
     }
   };
 
@@ -350,15 +321,24 @@ const connectNumericMenu: NumericMenuConnector = function connectNumericMenu(
         if (!connectorState.sendEvent) {
           connectorState.sendEvent = createSendEvent({
             instantSearchInstance,
-            helper,
-            attribute,
           });
+        }
+
+        const hasNoResults = results ? results.nbHits === 0 : true;
+        const preparedItems = prepareItems(state);
+        let allIsSelected = true;
+        for (const item of preparedItems) {
+          if (item.isRefined && decodeURI(item.value) !== '{}') {
+            allIsSelected = false;
+            break;
+          }
         }
 
         return {
           createURL: connectorState.createURL(state),
-          items: transformItems(prepareItems(state), { results }),
-          hasNoResults: results ? results.nbHits === 0 : true,
+          items: transformItems(preparedItems, { results }),
+          hasNoResults,
+          canRefine: !(hasNoResults && allIsSelected),
           refine: connectorState.refine,
           sendEvent: connectorState.sendEvent,
           widgetParams,