+++ src/types/utils.ts
@@ -18,3 +18,9 @@ export type RequiredKeys<TObject, TKeys extends keyof TObject> = Expand<
 >;
 
 export type Awaited<T> = T extends PromiseLike<infer U> ? Awaited<U> : T;
+
+/**
+ * Make certain keys of an object optional.
+ */
+export type PartialKeys<TObj, TKeys extends keyof TObj> = Omit<TObj, TKeys> &
+  Partial<Pick<TObj, TKeys>>;